radhesgar_blogs/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-nested-admin==4.1.1
djangorestframework==3.16.0
pillow==11.3.0
python-monkey-business==1.1.0
sqlparse==0.5.3
django-nested-admin==4.1.1
